redis: check the argument count for the missing-argument error, move wire tests to their own file
